pyproject.toml
# https://packaging.python.org/en/latest/specifications/#
# https://peps.python.org/pep-0621/
# https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html

# REM (not in venv shell): validate-pyproject pyproject.toml

[project]
# metadata-version = "2.3"
name = "pydtm-lib"
version = "0.0.3"
license = {text = "MIT"}
# summary = "Object model for standardising the structure of metadata applicable for mapping from one source data set to a target. Potentially being reshaped during the transfer."
description = "Object model for standardising the structure of metadata applicable for mapping from one source data set to a target. Potentially being reshaped during the transfer."
keywords = ["data pipelines", "data", "pipelines", "metadata"]
# https://pypi.org/classifiers/
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Development Status :: 2 - Pre-Alpha",
]
readme = "README.md"
# home-page = "https://github.com/scott-diprose/pydtm-lib"

requires-python = ">=3.10"
dependencies = [
    "jsonschema",
    "pyyaml",
    "lxml"
]

[project.optional-dependencies]
dev = [
    "safety",
    "pylint",
    "yapf",
    "pytest",
    "pytest-cov",
    "build",
    "twine"
]

# [project.scripts]
# pydtm_migrate = "pydtm.util.upgrade:migrate_version"


[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools >= 58.0",
    "wheel",
    # "setuptools-scm>=6.2",
]

# # https://setuptools.pypa.io/en/latest/userguide/package_discovery.html

[tool.setuptools.packages.find]
where = ["src"]
include = ["pydtm*"]
# include = ["pydtm", "pydtm.models", "pydtm.plugins*"]


[tool.pylint.MASTER]
# init-hook = "import sys; import os; sys.path.append('src')"
[tool.pylint.STRING]
check-quote-consistency = "yes"
[tool.pylint.'MESSAGES CONTROL']
max-line-length = 88
disable = [
    "missing-module-docstring",
    "no-member",
    # "too-few-public-methods",
]

[tool.yapf]
column_limit = 88

[tool.pytest.ini_options]

[tool.coverage.run]
source = ["src"]
